Set up the skeleton of a NotificationBuilder
MessageThreadNotifications are stupidly complex and I'm done trying to
construct them in-situ everywhere they're needed.
